use migrated collection hook
